refactor: Move parse_list to csvkit.cli
